fix(experiments): cast COUNT(*) subqueries to int to fix Dapper materialization (500 on /api/projects and /api/samples)
PostgreSQL COUNT(*) returns bigint (Int64) but ProjectListRow.BatchCount,
AnimalCount and SampleListRow.AnalysisCount were declared as int (Int32).
Dapper could not find a matching constructor and threw a 500. Added ::int
casts in both SQL subqueries so the wire type matches the C# record.

+++ src/Modules/Experiments/SISLAB.Modules.Experiments.Application/Biobank/Queries/ListSamplesQuery.cs
@@ -64,7 +64,7 @@ internal sealed class ListSamplesQueryHandler
                 COALESCE(c.consumed_value, 0)                        AS consumed_value,
                 s.collected_value - COALESCE(c.consumed_value, 0)    AS remaining_value,
                 s.collected_unit,
-                (SELECT COUNT(*) FROM experiments.sample_analyses a WHERE a.sample_id = s.id) AS analysis_count,
+                (SELECT COUNT(*) FROM experiments.sample_analyses a WHERE a.sample_id = s.id)::int AS analysis_count,
                 s.collected_at_utc,
                 ROW_NUMBER() OVER (ORDER BY s.collected_at_utc DESC, s.id DESC) AS row_number,
                 (COUNT(*) OVER ())::int AS total_rows
